# Optimize serial gemm + Fix parallel result