--- a/benchmarks/gemm/gemm_bench_float32.nim
+++ b/benchmarks/gemm/gemm_bench_float32.nim
@@ -54,7 +54,12 @@ const
   NbSamples = 10    # This might stresss the allocator when packing if the matrices are big
   CpuGhz = 3.6      # i9-9980XE OC All turbo 4.1GHz (AVX2 4.0GHz, AVX512 3.6GHz)
   NumCpuCores = 18
-  CpuFlopCycle = 64 # AVX2: 2xFMA/cycle = 2x8x2 - 2 x 8 floats x (1 add + 1 mul)
+  VectorWidth = 16  # 8 float32 for AVX2, 16 for AVX512
+  InstrCycle = 2    # How many instructions per cycle, (2xFMAs or 1xFMA for example)
+  FlopInstr = 2     # How many FLOP per instr (FMAs = 1 add + 1 mul)
+
+  TheoSerialPeak = CpuGhz * VectorWidth * InstrCycle * FlopInstr
+  TheoThreadedPeak = TheoSerialPeak * NumCpuCores
 
 const
   ashape: MatrixShape = (M, K)
@@ -220,8 +225,8 @@ when isMainModule:
   echo &"Required number of operations: {req_ops.float / float(10^6):>9.3f} millions"
   echo &"Required bytes:                {req_bytes.float / float(10^6):>9.3f} MB"
   echo &"Arithmetic intensity:          {req_ops.float / req_bytes.float:>9.3f} FLOP/byte"
-  echo &"Theoretical peak single-core:  {CpuGhz * CpuFlopCycle:>9.3f} GFLOP/s"
-  echo &"Theoretical peak multi:        {CpuGhz * CpuFlopCycle * NumCpuCores:>9.3f} GFLOP/s"
+  echo &"Theoretical peak single-core:  {TheoSerialPeak:>9.3f} GFLOP/s"
+  echo &"Theoretical peak multi:        {TheoThreadedPeak:>9.3f} GFLOP/s"
   echo "Make sure to not bench Apple Accelerate or the default Linux BLAS."
   block:
     let a = newSeqWith(M*K, float32 rand(1.0))