This optimizes the serial GEMM especially for AVX512.

It also fixes threading. The wrong part of A was retrieved on matrix borders.
The fix switched to OpenMP tasks instead of parallel for + taskloop.
I can't get the taskloop construct to not produce errors.

OpenBLAS version is 3.5 from Dec 31 2018, which provides a Skylake-X specific kernel:
  https://github.com/xianyi/OpenBLAS/blob/v0.3.5/kernel/x86_64/sgemm_kernel_16x4_skylakex.S

Serial speed is 1.5x OpenBLAS with AVX512 or 0.98x when using AVX2.
Unfortunately scaling is only about half the total number of core.

```
Openblas Serial
Collected 10 samples in 1.394 seconds
Average time: 138.927 ms
Stddev  time: 0.605 ms
Min     time: 138.614 ms
Max     time: 140.625 ms
Perf:         101.893 GFLOP/s

Laser AVX2 serial
Collected 10 samples in 1.491 seconds
Average time: 148.622 ms
Stddev  time: 0.897 ms
Min     time: 146.721 ms
Max     time: 149.429 ms
Perf:         95.247 GFLOP/s

Laser AVX512 serial
Collected 10 samples in 0.926 seconds
Average time: 92.050 ms
Stddev  time: 0.434 ms
Min     time: 91.587 ms
Max     time: 92.922 ms
Perf:         153.784 GFLOP/s

Openblas parallel
Collected 10 samples in 0.087 seconds
Average time: 8.037 ms
Stddev  time: 5.326 ms
Min     time: 6.325 ms
Max     time: 23.195 ms
Perf:         1761.421 GFLOP/s

Laser AVX2 parallel
Collected 10 samples in 0.126 seconds
Average time: 12.055 ms
Stddev  time: 4.081 ms
Min     time: 10.687 ms
Max     time: 23.670 ms
Perf:         1174.237 GFLOP/s

Laser AVX512 parallel
Collected 10 samples in 0.088 seconds
Average time: 8.262 ms
Stddev  time: 3.806 ms
Min     time: 6.921 ms
Max     time: 19.093 ms
Perf:         1713.268 GFLOP/s
```