@mratsim 2019-01-30
For reference

with 2 threads and AVX512
```
$  OMP_NUM_THREADS=2 ./build/gemm_f32_omp
Warmup: 0.9041 s, result 224 (displayed to avoid compiler optimizing warmup away)

A matrix shape: (M: 1920, N: 1920)
B matrix shape: (M: 1920, N: 1920)
Output shape: (M: 1920, N: 1920)
Required number of operations: 14155.776 millions
Required bytes:                   29.491 MB
Arithmetic intensity:            480.000 FLOP/byte
Theoretical peak single-core:    230.400 GFLOP/s
Theoretical peak multi:         4147.200 GFLOP/s
Make sure to not bench Apple Accelerate or the default Linux BLAS.

OpenBLAS benchmark
Collected 10 samples in 0.481 seconds
Average time: 47.604 ms
Stddev  time: 4.439 ms
Min     time: 46.016 ms
Max     time: 60.233 ms
Perf:         297.364 GFLOP/s

Laser production implementation
Collected 10 samples in 0.479 seconds
Average time: 47.373 ms
Stddev  time: 0.386 ms
Min     time: 47.001 ms
Max     time: 48.397 ms
Perf:         298.818 GFLOP/s
Mean Relative Error compared to vendor BLAS: 3.045843413929106e-06
```

with 4 threads and AVX512

```
$  OMP_NUM_THREADS=4 ./build/gemm_f32_omp
Warmup: 0.9030 s, result 224 (displayed to avoid compiler optimizing warmup away)

A matrix shape: (M: 1920, N: 1920)
B matrix shape: (M: 1920, N: 1920)
Output shape: (M: 1920, N: 1920)
Required number of operations: 14155.776 millions
Required bytes:                   29.491 MB
Arithmetic intensity:            480.000 FLOP/byte
Theoretical peak single-core:    230.400 GFLOP/s
Theoretical peak multi:         4147.200 GFLOP/s
Make sure to not bench Apple Accelerate or the default Linux BLAS.

OpenBLAS benchmark
Collected 10 samples in 0.287 seconds
Average time: 28.179 ms
Stddev  time: 3.451 ms
Min     time: 26.398 ms
Max     time: 37.837 ms
Perf:         502.350 GFLOP/s

Laser production implementation
Collected 10 samples in 0.258 seconds
Average time: 25.254 ms
Stddev  time: 0.060 ms
Min     time: 25.148 ms
Max     time: 25.381 ms
Perf:         560.530 GFLOP/s
Mean Relative Error compared to vendor BLAS: 3.045843413929106e-06
```