@mratsim 2019-01-30
Comparison with various problem size

120x120
```
A matrix shape: (M: 120, N: 120)
B matrix shape: (M: 120, N: 120)
Output shape: (M: 120, N: 120)
Required number of operations:     3.456 millions
Required bytes:                    0.115 MB
Arithmetic intensity:             30.000 FLOP/byte
Theoretical peak single-core:    230.400 GFLOP/s
Theoretical peak multi:         4147.200 GFLOP/s
Make sure to not bench Apple Accelerate or the default Linux BLAS.

OpenBLAS benchmark
Collected 10000 samples in 0.292 seconds
Average time: 0.026 ms
Stddev  time: 0.240 ms
Min     time: 0.020 ms
Max     time: 22.340 ms
Perf:         133.374 GFLOP/s

Laser production implementation
Collected 10000 samples in 0.647 seconds
Average time: 0.064 ms
Stddev  time: 0.190 ms
Min     time: 0.056 ms
Max     time: 12.184 ms
Perf:         54.081 GFLOP/s
Mean Relative Error compared to vendor BLAS: 0.0
```

240x240
```
A matrix shape: (M: 240, N: 240)
B matrix shape: (M: 240, N: 240)
Output shape: (M: 240, N: 240)
Required number of operations:    27.648 millions
Required bytes:                    0.461 MB
Arithmetic intensity:             60.000 FLOP/byte
Theoretical peak single-core:    230.400 GFLOP/s
Theoretical peak multi:         4147.200 GFLOP/s
Make sure to not bench Apple Accelerate or the default Linux BLAS.

OpenBLAS benchmark
Collected 10000 samples in 0.690 seconds
Average time: 0.055 ms
Stddev  time: 0.240 ms
Min     time: 0.050 ms
Max     time: 22.902 ms
Perf:         502.465 GFLOP/s

Laser production implementation
Collected 10000 samples in 2.026 seconds
Average time: 0.187 ms
Stddev  time: 0.301 ms
Min     time: 0.109 ms
Max     time: 19.423 ms
Perf:         147.762 GFLOP/s
Mean Relative Error compared to vendor BLAS: 0.0
```

360x360
```
A matrix shape: (M: 360, N: 360)
B matrix shape: (M: 360, N: 360)
Output shape: (M: 360, N: 360)
Required number of operations:    93.312 millions
Required bytes:                    1.037 MB
Arithmetic intensity:             90.000 FLOP/byte
Theoretical peak single-core:    230.400 GFLOP/s
Theoretical peak multi:         4147.200 GFLOP/s
Make sure to not bench Apple Accelerate or the default Linux BLAS.

OpenBLAS benchmark
Collected 10000 samples in 2.033 seconds
Average time: 0.176 ms
Stddev  time: 0.340 ms
Min     time: 0.158 ms
Max     time: 18.097 ms
Perf:         531.223 GFLOP/s

Laser production implementation
Collected 10000 samples in 3.038 seconds
Average time: 0.270 ms
Stddev  time: 0.315 ms
Min     time: 0.185 ms
Max     time: 19.048 ms
Perf:         345.651 GFLOP/s
Mean Relative Error compared to vendor BLAS: 0.0
```

480x480
```
Warmup: 0.9046 s, result 224 (displayed to avoid compiler optimizing warmup away)

A matrix shape: (M: 480, N: 480)
B matrix shape: (M: 480, N: 480)
Output shape: (M: 480, N: 480)
Required number of operations:   221.184 millions
Required bytes:                    1.843 MB
Arithmetic intensity:            120.000 FLOP/byte
Theoretical peak single-core:    230.400 GFLOP/s
Theoretical peak multi:         4147.200 GFLOP/s
Make sure to not bench Apple Accelerate or the default Linux BLAS.

OpenBLAS benchmark
Collected 10000 samples in 2.648 seconds
Average time: 0.206 ms
Stddev  time: 0.305 ms
Min     time: 0.177 ms
Max     time: 22.654 ms
Perf:         1074.502 GFLOP/s

Laser production implementation
Collected 10000 samples in 4.176 seconds
Average time: 0.359 ms
Stddev  time: 0.333 ms
Min     time: 0.225 ms
Max     time: 19.364 ms
Perf:         615.330 GFLOP/s
Mean Relative Error compared to vendor BLAS: 2.02139403882029e-06
```